requirements.txt
Flask==2.3.3
pycryptodomex==4.18.0
bcrypt==4.0.1
requests==2.31.0
